tests/wpt/mozilla/tests/bluetooth/getCharacteristics/blocklisted-characteristics.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.blocklist);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [device_information.name]}]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(device_information.name))
    .then(service => promise_rejects_dom(t, 'NotFoundError', service.getCharacteristics()));
}, 'The Device Information service is composed of blocklisted characteristics so we shouldn\'t find any.');
</script>
